What? Another planet and a red line?
Yes, I added Venus because I'm feeling ethically obligated and a new... red line! 
This red line represents the trayectory that the shuttle must go through, so you better click at the moment mars crosses the line.
You don't have to do it yet because clicking will be a minor but a yet waste of your energy, it's not implemented yet.
Pressing the 'd' makes fun stuff though.